+++ qml/DecoDXLog/PropagationPanel.qml
@@ -97,6 +97,7 @@ Item {
         anchors.fill: parent
         contentWidth: availableWidth
         clip: true
+        ScrollBar.vertical: PanelScrollBar {}
 
         ColumnLayout {
             width: parent.width